+++ src/interfaces/FillEvent.ts
@@ -1,0 +1,12 @@
+export type FillEvent = {
+  id: string;
+  userId: string;
+  cylinderSetId: string;
+  cylinderSetName: string;
+  gasMixture: string;
+  description: string;
+  price: number;
+  createdAt: string;
+};
+
+export default FillEvent;
