+++ src/lib/apiRequests/fillEventRequests.ts
@@ -1,15 +1,5 @@
 import { authGetAsync } from './api';
-
-export type FillEvent = {
-  id: string;
-  userId: string;
-  cylinderSetId: string;
-  cylinderSetName: string;
-  gasMixture: string;
-  description: string;
-  price: number;
-  createdAt: string;
-};
+import FillEvent from '../../interfaces/FillEvent';
 
 export const getFillEvents = async (): Promise<FillEvent[]> => {
   const response = await authGetAsync<FillEvent[]>(
